refactor: command parser
